% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Candidate.R
\docType{class}
\name{Candidate-class}
\alias{Candidate-class}
\alias{initialize,Candidate-method}
\alias{show,Candidate-method}
\title{A Candidate object}
\description{
Objects of class \code{Candidate} are created by the \code{addCandidate} function. Objects of
this class include \code{show} and \code{print} methods.
}
\details{
An object of the class `Candidate' has the following slots:
\itemize{
\item \code{name} Name of the presidential candidate (character string).
\item \code{delegatesWon} Number of delegates won by given candidate (numeric).
\item \code{party} Party of given candidate, 'Republican' or 'Democratic' (character string).
\item \code{delegatesNeeded} Number of delegates needed by given candidate (numeric).
}
}
\author{
Jeff Ziegler
}

