% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createRace.R
\name{createRace}
\alias{createRace}
\alias{createRace,ANY-method}
\title{Create an object of class 'Candidate'}
\usage{
createRace(allCandidates, party)
}
\arguments{
\item{party}{Party of given candidate, 'Republican' or 'Democratic' (character string).}

\item{name}{Name of the presidential candidate (character string).}

\item{delegatesWon}{Number of delegates won by given candidate (numeric).}
}
\value{
An object of class `Candidate' containing
\itemize{
\item allCandidates Object(s) of class 'Candidate'. Candidates should all 
be from the same political party (list).
\item party Party of candidates, 'Republican' or 'Democratic' (character string).
\item delegatesNeeded Number of delegates needed by given candidate (numeric).
Automatically determined when method is initialized with creation of a new instance of the class 'Race'.
\item delegatesRemaining Number of delegates not yet allocated in a given party primary (numeric).
Automatically determined when method is initialized with creation of a new instance of the class 'Race'.
}
}
\description{
Function creates an object of class `Candidate' and determines the number of
delegates the candidate needs to win their party's presidential nomination.
}
\examples{

clinton <- createCandidate("H. Clinton", 1614, "Democrat")
sanders <- createCandidate("B. Sanders", 856, "Democrat")
demCandidates <- list(clinton, sanders)
demRace <- createRace(allCandidates = demCandidates, party = "Republican")
plot(x = demRace)

}
\author{
Jeff Ziegler
}
\seealso{
\code{\link{Race}}
}

