\name{primaryRacePack-package}
\alias{primaryRacePack-package}
\alias{primaryRacePack}
\docType{package}
\title{
\packageTitle{primaryRacePack}
}
\description{
\packageDescription{primaryRacePack}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{primaryRacePack}
\packageIndices{primaryRacePack}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{primaryRacePack}

Maintainer: \packageMaintainer{primaryRacePack}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
