% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primaryRacePack-package.R
\docType{package}
\name{primaryRacePack}
\alias{primaryRacePack}
\alias{primaryRacePack-package}
\title{primaryRacePack}
\description{
The primaryRacePack package calculates the number of delegates needed based on the number of delegates won and remaining.
}
\examples{

newCandidate <- createCandidate("Ben Carson", 8, "Republican")
str(newCandidate)
propNeeded(newCandidate, 1396)

}
\author{
Jeff Ziegler: \email{jeffreyziegler@wustl.edu}
}
\seealso{
\code{\link{createCandidate}}, \code{\link{Candidate}}, \code{\link{propNeeded}}

\dontrun{
demo(createCandidate.R)
demo(propNeeded.R)
}
}

