% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/propNeeded.R
\name{propNeeded}
\alias{propNeeded}
\alias{propNeeded,ANY-method}
\title{Calculate the proportion of delegates a given candidate needs to win nomination}
\usage{
propNeeded(object, remainingDelegates = NULL)
}
\arguments{
\item{object}{An object of the class 'Candidate'.}

\item{remainingDelegates}{Number of delegates not yet allocated in a given candidate's party primary (numeric).}
}
\value{
Proportion of the delegates not yet allocated that a given candidate needs to win party's nomination (numeric).
}
\description{
Function determines the proportion of remaining delegates a presidential candidate needs 
to secure in order to win his party's nomination.
}
\examples{

newCandidate <- createCandidate("John Kasich", 143, "Republican")
propNeeded(newCandidate)
propNeeded(newCandidate, 1396) 

}
\author{
Jeff Ziegler
}
\seealso{
\code{\link{createCandidate}}, \code{\link{Candidate}}
}

