% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createCandidate.R
\name{createCandidate}
\alias{createCandidate}
\alias{createCandidate,ANY-method}
\title{Create an object of class 'Candidate'}
\usage{
createCandidate(name, delegatesWon, party)
}
\arguments{
\item{name}{Name of the presidential candidate (character string).}

\item{delegatesWon}{Number of delegates won by given candidate (numeric).}

\item{party}{Party of given candidate, 'Republican' or 'Democratic' (character string).}
}
\value{
An object of class `Candidate' containing
\itemize{
\item name Name of the presidential candidate (character string).
\item delegatesWon Number of delegates won by given candidate (numeric).
\item party Party of given candidate, 'Republican' or 'Democratic' (character string).
\item{delegatesNeeded} Number of delegates needed by given candidate to obtain nomination (numeric).
Automatically determined when method is initialized with creation of a new instance of the class 'Candidate'.
 }
}
\description{
Function creates an object of class `Candidate' and determines the number of
delegates the candidate needs to win their party's presidential nomination.
}
\examples{

newCandidate <-createCandidate("Martin O'Malley", 0, "Democratic") 
newCandidate <- createCandidate("Ben Carson", 8, "Republican") 

}
\author{
Jeff Ziegler
}
\seealso{
\code{\link{Candidate}}
}

